#23 Add Write Prediction Results To Database
This commit adds a function to take the prediction results from a job_batch, extract the values for the fields that need to be updated, and writes these updates to the Requests table for all the processed jobs.

+++ server/application/models.py
@@ -157,7 +157,7 @@ class Users(AbstractUser):
 class Requests(models.Model):
     request_id = models.AutoField(primary_key=True)
     created_at = models.IntegerField(null=False, blank=False)
-    probability = models.IntegerField(blank=True, null=True)
+    probability = models.FloatField(blank=True, null=True)
     image = models.BinaryField(blank=False, null=False)
 
     LOCALIZATION_CHOICES = [
